Change Rectangle intersection behavior and add tests
Now when intersecting a rectangle with another that's non-overlapping, the
result contains useful information about how much it does overlap along each
axis and the closest edge or corner. This makes tile intersection work as
desired when min_overlap is non-zero and the tiles just touch or don't overlap
at all.

Also added the beginning of a pytest-compatible test suite, starting with
a test for Rectangle.intersection.